coap: Split into version independent and versioned parts
