+++ Cargo.toml
@@ -39,7 +39,7 @@ coap-numbers = "^0.2.0"
 
 embedded-graphics = "0.6"
 
-coap-message = "^0.2.3"
+coap-message-0-2 = { package = "coap-message", version = "^0.2.3" }
 coap-handler = "^0.1.4"
 embedded-nal = { version = "0.6.0", optional = true }
 embedded-nal-tcpextensions = { version = "0.1", optional = true }
